OS is a global const set at init, passed to Dash constructor.
Assuming that the OS will not change every is reasonable, and the OS is
passed to the Dash constructor, which sets it as well as the Dash URI
handler as private variables. We can technically skip passing the OS to
the constructor, but that would make it harder to debug for multiple
platforms.